pyproject.toml
[project]
name = "multi-agent-langgraph"
version = "0.1.0"
description = "Multi-Agent Document Analysis with LangGraph - Workshop by learnwithparam.com"
authors = [{ name = "Param Harrison", email = "param@learnwithparam.com" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.6.0",
    "httpx>=0.27.0",
    "openai>=1.0.0",
    "google-generativeai>=0.8.0",
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-google-genai>=2.0.0",
    "pymupdf4llm>=0.0.10",
    "python-multipart>=0.0.6",
    "python-docx>=1.0.0",
    "llama-index-core>=0.10.0",
    "numpy>=1.24.0",
    "aiohttp>=3.9.0",
]
